+++ src/utils/re_api.py
@@ -10,7 +10,19 @@ _CONF = get_config()
 
 
 def query(name, params, tok=None):
-    """Run a stored query from the RE API."""
+    """
+    Run a stored query from the RE API.
+
+    Returns (from relation_engine)
+    {
+        "results": list,        # array of result docs, might be empty
+        "count": int,           # num docs
+        "has_more": bool,       # whether more results for cursor on server
+        "cursor_id": str/None,  # id of temporary cursor on server
+        "stats": dict,          # stats
+    }
+
+    """
     resp = requests.post(
         _CONF['re_url'] + '/api/v1/query_results',
         params={'stored_query': name},